+++ src/deploy/creation_common.rs
@@ -284,6 +284,7 @@ pub(super) mod dependencies {
             .display_order(DisplayOrder::Dependencies as usize)
     }
 
+    #[allow(clippy::iter_overeager_cloned)]
     pub(in crate::deploy) fn get<'a>(matches: &'a ArgMatches) -> Vec<&'a str> {
         matches
             .values_of(ARG_NAME)
@@ -373,6 +374,7 @@ pub(super) mod arg_simple {
                 .requires(super::session::ARG_NAME)
         }
 
+        #[allow(clippy::iter_overeager_cloned)]
         pub fn get<'a>(matches: &'a ArgMatches) -> Vec<&'a str> {
             matches
                 .values_of(ARG_NAME)
@@ -393,6 +395,7 @@ pub(super) mod arg_simple {
                 .requires(super::payment::ARG_NAME)
         }
 
+        #[allow(clippy::iter_overeager_cloned)]
         pub fn get<'a>(matches: &'a ArgMatches) -> Vec<&'a str> {
             matches
                 .values_of(ARG_NAME)